[FIX] Remove Claim Emblems and Fix Banner Purchased Tests
