+++ src/features/game/events/landExpansion/bannerPurchased.ts
@@ -22,33 +22,20 @@ type Options = {
   farmId?: number;
 };
 
-export const BANNER_PRICE_INCREASED_DATE = new Date("2024-08-01T00:00:00Z");
-
 export function getBannerPrice(
   banner: SeasonalBanner | "Lifetime Farmer Banner",
   hasPreviousBanner: boolean,
   hasLifetimeBanner: boolean,
-  hasGoldPass: boolean,
   createdAt: number,
   farmId?: number,
 ): Decimal {
-  const bannerPriceIncreased =
-    new Date(createdAt).getTime() > BANNER_PRICE_INCREASED_DATE.getTime();
-  const lifeTimePrice = bannerPriceIncreased ? 740 : 540;
-  const seasonBannerIncrease = bannerPriceIncreased ? 10 : 0;
+  const lifeTimePrice = 740;
 
   if (banner === "Lifetime Farmer Banner") {
     return new Decimal(lifeTimePrice);
   }
 
-  const goldPassRetired =
-    new Date(createdAt).getTime() > new Date("2024-08-01T00:00:00Z").getTime();
-  const getsFreeBanner =
-    !goldPassRetired &&
-    farmId &&
-    GOLD_PASS_LESS_THAN_THREE_MONTHS.includes(farmId);
-
-  if (hasLifetimeBanner || getsFreeBanner) return new Decimal(0);
+  if (hasLifetimeBanner) return new Decimal(0);
 
   const season = getSeasonByBanner(banner);
   const seasonStartDate = SEASONS[season].startDate;
@@ -59,22 +46,17 @@ export function getBannerPrice(
     (createdAt - seasonStartDate.getTime()) / WEEK,
   );
 
-  const goldPassDiscount = hasGoldPass && !goldPassRetired ? 15 : 0;
-
   if (weeksElapsed < 2) {
     const previousBannerDiscount = hasPreviousBanner ? 15 : 0;
-    return new Decimal(65)
-      .add(seasonBannerIncrease)
-      .sub(previousBannerDiscount)
-      .sub(goldPassDiscount);
+    return new Decimal(75).sub(previousBannerDiscount);
   }
   if (weeksElapsed < 4) {
-    return new Decimal(90).add(seasonBannerIncrease).sub(goldPassDiscount);
+    return new Decimal(100);
   }
   if (weeksElapsed < 8) {
-    return new Decimal(70).add(seasonBannerIncrease).sub(goldPassDiscount);
+    return new Decimal(80);
   }
-  return new Decimal(50).add(seasonBannerIncrease).sub(goldPassDiscount);
+  return new Decimal(60);
 }
 
 export function purchaseBanner({
@@ -97,10 +79,7 @@ export function purchaseBanner({
       throw new Error("You already have this banner");
     }
 
-    const lifeTimerFarmerPriceIncreased =
-      new Date(createdAt).getTime() >
-      new Date("2024-08-01T00:00:00Z").getTime();
-    const lifeTimePrice = lifeTimerFarmerPriceIncreased ? 740 : 540;
+    const lifeTimePrice = 740;
 
     if (currentBlockBucks.lessThan(lifeTimePrice)) {
       throw new Error("Insufficient Block Bucks");
@@ -132,13 +111,11 @@ export function purchaseBanner({
   const hasLifetimeBanner = (
     inventory["Lifetime Farmer Banner"] ?? new Decimal(0)
   ).gt(0);
-  const hasGoldPass = (inventory["Gold Pass"] ?? new Decimal(0)).gt(0);
 
   const price = getBannerPrice(
     action.name,
     hasPreviousBanner,
     hasLifetimeBanner,
-    hasGoldPass,
     createdAt,
     farmId,
   );
@@ -152,742 +129,3 @@ export function purchaseBanner({
 
   return stateCopy;
 }
-
-const GOLD_PASS_LESS_THAN_THREE_MONTHS = [
-  25, 168, 249, 610, 717, 1119, 1206, 1519, 1584, 1600, 1644, 1701, 2456, 2640,
-  2847, 2889, 3123, 3541, 3646, 3685, 3820, 3834, 4236, 5394, 5704, 5905, 6315,
-  6866, 7011, 7086, 7562, 8111, 8238, 8391, 8502, 9221, 9775, 10009, 10252,
-  10581, 10626, 11374, 11415, 12185, 12615, 15321, 15360, 15369, 15493, 15850,
-  15916, 16016, 17043, 17410, 18337, 19317, 19340, 19403, 20401, 20610, 21686,
-  21766, 21953, 23141, 25366, 25699, 25888, 25964, 28095, 29145, 30011, 30434,
-  30854, 32364, 33747, 39109, 40739, 40760, 41613, 42247, 42502, 42729, 44851,
-  44981, 47294, 47489, 47648, 47900, 47995, 48395, 49806, 52163, 52214, 54860,
-  54989, 56187, 56424, 56510, 56685, 58041, 59028, 62561, 64620, 65001, 65145,
-  67565, 70068, 72403, 75280, 78636, 80431, 82122, 83585, 85634, 89092, 90118,
-  92566, 95379, 99643, 100183, 100226, 100953, 104044, 108523, 110050, 110143,
-  113171, 118789, 119715, 123008, 123595, 124105, 125783, 126898, 128628,
-  129088, 129985, 130739, 132452, 133042, 133507, 136988, 137191, 137663,
-  139269, 139807, 140598, 141047, 141067, 141488, 141558, 141753, 142357,
-  143759, 143924, 144666, 146717, 147050, 147483, 148381, 149388, 150007,
-  150452, 150766, 151298, 151364, 151726, 152374, 152450, 152508, 152601,
-  152847, 152877, 153204, 153244, 154413, 154789, 155171, 155443, 155857,
-  156299, 156753, 156915, 157462, 158704, 159937, 159971, 160608, 160737,
-  161004, 161074, 163850, 164056, 164436, 164694, 165052, 165063, 165261,
-  165361, 165464, 165483, 165544, 165688, 165754, 165762, 165774, 165896,
-  165906, 165910, 165911, 165935, 165948, 165955, 166034, 166079, 166101,
-  166151, 166227, 166305, 166324, 166373, 166386, 166390, 166397, 166439,
-  166457, 166487, 166552, 166567, 166579, 166637, 166667, 166670, 166699,
-  166818, 167250, 167330, 167331, 167493, 167585, 167588, 167732, 167751,
-  167756, 167798, 168127, 168186, 168202, 168340, 168347, 169188, 169320,
-  169328, 169349, 169358, 169437, 169469, 169596, 169610, 169754, 169869,
-  169994, 170045, 170053, 170096, 170341, 170968, 170987, 171039, 171246,
-  171556, 171648, 171716, 171894, 171978, 172090, 172104, 172438, 172597,
-  172617, 172674, 172790, 173078, 173198, 173558, 173735, 173738, 174045,
-  174090, 174102, 174250, 174276, 174290, 174352, 174491, 174585, 174593,
-  174845, 174878, 175220, 175243, 175297, 175951, 175955, 176154, 176275,
-  176305, 176310, 176364, 176374, 176380, 176436, 176751, 177372, 177432,
-  177453, 177561, 177802, 177942, 177967, 178503, 178979, 178983, 178984,
-  179014, 179024, 179227, 179258, 179294, 179407, 180039, 180054, 180126,
-  180269, 180768, 180859, 180863, 180903, 180920, 180969, 180970, 180981,
-  181104, 181128, 181282, 181324, 181407, 181460, 181502, 181737, 181762,
-  181765, 181767, 182103, 182195, 182253, 182409, 182729, 182813, 182888,
-  183179, 184036, 184126, 184353, 184565, 184623, 184653, 184826, 184881,
-  184922, 185026, 185027, 185494, 185531, 185714, 185732, 185734, 185966,
-  186225, 186480, 186705, 186717, 186725, 186777, 187038, 187060, 187092,
-  187620, 187787, 188144, 188197, 188296, 188417, 188482, 188550, 188609,
-  188965, 189259, 189295, 189356, 189374, 189591, 189752, 189759, 189801,
-  190158, 190177, 190236, 190773, 191051, 191131, 191249, 191464, 191591,
-  191731, 191765, 191777, 191834, 191840, 191909, 191910, 191961, 192009,
-  192078, 192154, 192162, 192337, 192447, 192480, 192509, 192623, 192625,
-  192739, 192811, 192812, 193030, 193099, 193146, 193171, 193191, 193245,
-  193383, 193463, 193546, 193654, 193707, 193772, 193775, 193825, 193925,
-  194106, 194117, 194143, 194242, 194363, 194728, 194731, 194749, 194912,
-  194957, 194979, 195010, 195507, 195589, 195715, 195723, 195817, 195855,
-  195951, 196010, 196024, 196043, 196052, 196063, 196064, 196082, 196085,
-  196108, 196109, 196111, 196177, 196268, 196324, 196464, 196585, 196674,
-  196710, 196872, 196975, 197010, 197075, 197248, 197265, 197274, 197299,
-  197301, 197302, 197338, 197370, 197375, 197397, 197406, 197420, 197431,
-  197557, 197606, 197613, 197616, 197624, 197639, 197649, 197653, 197656,
-  197663, 198034, 198046, 198270, 198322, 198326, 198343, 198361, 198371,
-  198373, 198488, 198510, 198635, 198764, 199382, 199606, 199618, 199791,
-  199862, 200030, 200039, 200067, 200173, 200175, 200177, 200178, 200247,
-  200429, 200483, 200508, 200520, 200528, 200539, 200593, 200708, 200711,
-  200717, 200725, 200730, 200801, 200825, 200859, 200946, 200947, 200978,
-  201020, 201023, 201031, 201041, 201057, 201097, 201150, 201151, 201163,
-  201173, 201215, 201251, 201380, 201465, 201466, 201467, 201703, 201774,
-  201858, 201877, 201882, 201886, 201896, 201899, 201903, 201911, 201919,
-  201992, 202017, 202031, 202037, 202049, 202050, 202247, 202262, 202269,
-  202296, 202297, 202327, 202382, 202411, 202413, 202417, 202453, 202459,
-  202491, 202510, 202527, 202544, 202569, 202577, 202585, 202586, 202656,
-  202682, 202758, 202909, 202940, 202952, 202964, 203054, 203130, 203143,
-  203184, 203196, 203209, 203212, 203338, 203400, 203435, 203442, 203540,
-  203568, 203593, 203623, 203681, 203699, 203750, 203753, 203776, 203798,
-  203825, 203831, 203864, 203948, 203967, 203987, 204036, 204163, 204169,
-  204295, 204300, 204301, 204399, 204452, 204468, 204484, 204539, 204628,
-  204635, 204647, 204661, 204667, 204687, 204692, 204703, 204720, 204721,
-  204723, 204726, 204727, 204728, 204734, 204736, 204738, 204767, 204768,
-  204769, 204770, 204771, 204772, 204780, 204787, 204806, 204814, 204831,
-  204833, 204834, 204835, 204838, 204839, 204841, 204843, 204898, 204912,
-  204948, 204977, 204982, 204983, 204986, 204995, 204997, 205005, 205084,
-  205101, 205135, 205156, 205184, 205186, 205213, 205216, 205243, 205285,
-  205384, 205390, 205394, 205447, 205538, 205580, 205630, 205652, 205669,
-  205692, 205697, 205724, 205731, 205750, 205789, 205834, 205886, 205893,
-  205922, 205923, 205960, 206008, 206080, 206089, 206146, 206155, 206169,
-  206174, 206181, 206202, 206207, 206235, 206249, 206275, 206305, 206333,
-  206344, 206384, 206399, 206458, 206502, 206527, 206549, 206564, 206597,
-  206612, 206639, 206688, 206696, 206706, 206716, 206728, 206736, 206748,
-  206756, 206762, 206768, 206790, 206793, 206806, 206852, 206876, 206900,
-  206901, 206934, 206955, 206973, 207002, 207016, 207024, 207029, 3975493785964,
-  4880227969362, 4993034568832, 9051418630117, 11136236300452, 14197810638193,
-  14790429591423, 19037101976687, 22951303701236, 30672260576195,
-  32309167835276, 32396148965561, 32667073778986, 37201614492829,
-  37785185085250, 48725078719767, 48910960397058, 55505529163982,
-  57933660353869, 73722039945769, 76671106774079, 77082733836659,
-  81356949530304, 81760194183387, 83016838297821, 84371908531206,
-  86088288125125, 88325704800706, 93134237473966, 94024237823176,
-  100200029483485, 101048958409995, 113656721671946, 114039120392575,
-  116619998423390, 118304172738657, 118854109208172, 121164215174133,
-  121164982209348, 122162170930428, 129782400340778, 135166747330493,
-  141392144495937, 143666660282219, 150010191278544, 152806194327767,
-  153581605417395, 154503989121636, 156302399359476, 162382265037654,
-  165943784377509, 169554885156117, 172038604239254, 173579588093287,
-  179155485852220, 179781071521184, 185139279573611, 188271640192775,
-  189317743760318, 189340553175674, 193749628098018, 194639652619526,
-  198952863997293, 203184673921074, 203754642489910, 204800996782660,
-  207037773011552, 219045738621424, 225825059233348, 232326917902914,
-  234287706075175, 234959297105409, 235482393565912, 237665782619931,
-  237716433581942, 245413614392355, 247065824360431, 252817181505992,
-  254496967920548, 258999104723324, 261156553741425, 261693987397296,
-  263070881162414, 266476651544894, 267697342242116, 277714777214669,
-  281428994217720, 284541078330035, 285566523683430, 292264380574572,
-  294337173707582, 294975854197958, 300451573223909, 300501516472687,
-  306193474096278, 308299337664206, 310407449529831, 310521647772008,
-  316150974022132, 317218031602860, 321320931740093, 327090097967255,
-  332613001276429, 334856483357519, 336279036366509, 339668008442540,
-  343214932098736, 343294823771811, 346322309231904, 347625927439479,
-  355354770031497, 356266550414013, 365200874539432, 365491813992903,
-  367633778751277, 367891416193760, 368878856134130, 370105025483114,
-  374382075749212, 374842712589359, 394991618409689, 407448237307414,
-  408932166115027, 412159401607774, 413499773003275, 415527671164705,
-  415823759059664, 422719457531414, 425558998808909, 425682956807741,
-  431424165124689, 432202754089140, 435922962389490, 440370011419268,
-  442941021525721, 448595386830988, 452755001312642, 463070011187978,
-  471331198814934, 475259191954789, 475952625845562, 481107268614322,
-  484530205690069, 486238278503451, 490080226154671, 490245936628096,
-  490953381430928, 492172262645431, 499408171044089, 508888055013027,
-  510694120312525, 517205535024361, 520477690470921, 521462456865797,
-  527384070974340, 528309345483694, 530556929117658, 531374891636396,
-  533338529648320, 536609244300574, 536980614269734, 537545569229635,
-  539763561350586, 549903156999133, 550476825387807, 554279874695216,
-  566232484464813, 569788936375838, 571256488343611, 571512803265460,
-  573491687096218, 574004246416288, 577609205096570, 579734164405476,
-  582738499972259, 584546126740226, 584639032092190, 585643893951124,
-  586114991983337, 592458662309850, 595611360455249, 597583893210389,
-  598808045337775, 601029656504953, 606453958275044, 611522180013151,
-  614253347536008, 615144432315808, 621062546546941, 622390906097575,
-  628659125241921, 635772845169350, 635831057356171, 636360383724094,
-  637385707143200, 637535189671288, 640022174190417, 644566563046427,
-  650749989434066, 660587687483618, 661620346184902, 662450165875206,
-  666028720864702, 681517640215529, 692093980388161, 693496151938988,
-  701292348426493, 701959258589413, 714940245904089, 727581756521704,
-  728429127364313, 728626679051984, 735043309809039, 745434158884339,
-  757272893326157, 763369198175969, 778401762341929, 779115684077828,
-  779633413113468, 782970359707795, 784461017493132, 787097943100145,
-  792088757358469, 792514110160811, 793940022313119, 796200203954503,
-  797267900595527, 803663182421448, 805899641574509, 808354868186102,
-  810572934863797, 810772242101421, 814660228699336, 815070353919749,
-  815164585927181, 816188896940304, 816581300134756, 819189464602666,
-  823360808800950, 829640039996777, 834361257477038, 843859036424241,
-  845534708458044, 846052620104658, 846555721030409, 846936488686717,
-  849186239603292, 852483436405722, 859554098076603, 861020753588257,
-  861115902154422, 866476596940209, 869701689853206, 870284881593662,
-  876245094105224, 877362066819257, 904146165802836, 905396914427705,
-  906245046075290, 909083080178409, 909843494638866, 919790372665794,
-  927109409162500, 933312500065430, 939864196131770, 945563672479335,
-  947111206338288, 952844148717157, 954702516238427, 955862635473354,
-  963810068497724, 967147798373223, 969937402088329, 970141941662306,
-  972282465103359, 974709380777457, 976785889714434, 990452331282632,
-  990640382400888, 991501730514997, 991936500297360, 993825701847661,
-  994679967286630, 995911122597454, 999254974922260, 1000632219461761,
-  1003799945619473, 1004942037390963, 1006097509514382, 1010125660469023,
-  1015160993801549, 1015370103677471, 1018535253755829, 1020827251686130,
-  1021364045162156, 1025193182811658, 1030079686515283, 1030779775349710,
-  1033376959507022, 1038549956521038, 1038922610881741, 1041604686411699,
-  1050387853194942, 1050397760882694, 1051844501168187, 1052772810315886,
-  1053522646036294, 1053574123824324, 1054675457007978, 1057410771906211,
-  1060014359232835, 1060535917451130, 1061680807995818, 1062934285072714,
-  1065252427918882, 1067907158548532, 1068278300673865, 1070514453627298,
-  1071949435622375, 1076387307082120, 1077320515449937, 1083181477468064,
-  1083265446999263, 1083935425826194, 1084592065194363, 1090870586002785,
-  1091056159953636, 1092689453456966, 1098917088912807, 1101920246786704,
-  1103817587113671, 1111037986913915, 1119531061240275, 1123942054587038,
-  1126354294504055, 1128976301583508, 1131156523365541, 1135864956404118,
-  1137164551371754, 1141608347616651, 1144670514303248, 1154067054498057,
-  1155545343304484, 1157291187037392, 1159097886589904, 1162674853705985,
-  1165696599638776, 1165795865661737, 1166140193390145, 1169191736908583,
-  1179057307332857, 1181127083448696, 1185728982944412, 1188136289109469,
-  1188425212656944, 1188492243493157, 1190782813886454, 1190872732330479,
-  1191351882563385, 1194359512593119, 1194626999468774, 1196582962147320,
-  1202770754502503, 1208117374024119, 1211138349734457, 1212462898371363,
-  1221325664064040, 1223154677524849, 1234383767884861, 1240974172160747,
-  1243491043476958, 1246522192551083, 1249934267607881, 1252231467237460,
-  1257378602084960, 1264769056867766, 1267608468485436, 1269629456641935,
-  1270293928605726, 1270449467379234, 1273349091882209, 1273456820386463,
-  1274634274667933, 1278071759172925, 1278920260339927, 1284719152843318,
-  1289752073051512, 1291488522307639, 1291974004121636, 1292941298428915,
-  1293113060260277, 1295830734363059, 1302811617667579, 1304163407681544,
-  1307511184586448, 1311661074396652, 1312244041580063, 1321213243693712,
-  1328801281094674, 1335627043314858, 1336682354659311, 1339067821834851,
-  1346596327803703, 1347649167316962, 1350302543511402, 1352547540848482,
-  1361260780825034, 1363035182088003, 1363578735896555, 1368633586767766,
-  1372502051219164, 1376486851027978, 1378609912279441, 1381308626984431,
-  1384527879370616, 1388994991233671, 1392377786056928, 1392872081605704,
-  1395189001467422, 1400835042470217, 1416532609463619, 1421106378734129,
-  1423003006227343, 1427521658207856, 1429531714871154, 1430126138560102,
-  1439334966929193, 1441019408970544, 1445970558690696, 1449513021038868,
-  1450959575700406, 1455392350329811, 1465990333981782, 1466729547888111,
-  1467165695594789, 1469869937477638, 1472971440442394, 1473123078258114,
-  1475451559840199, 1485371896147916, 1491400709314273, 1498926374071923,
-  1502466187389360, 1503427419033452, 1506601286192130, 1510724232194341,
-  1511244458359199, 1513287415790795, 1522027538792054, 1528101311405550,
-  1534075179129915, 1538171683141182, 1542034844246917, 1544068064987598,
-  1545914496993121, 1554752408256347, 1558312565569358, 1581039221250660,
-  1594491255571892, 1596296758750219, 1608250181498375, 1614120976297325,
-  1614948312552767, 1626528876507774, 1629160809318440, 1634265842410606,
-  1637989060900805, 1642973110298093, 1647406123016767, 1653360300780171,
-  1654723171340508, 1661964325367792, 1666208992666851, 1666981470117319,
-  1667921427092391, 1671454406654975, 1676869107164127, 1693281832407226,
-  1693737448426490, 1696338607985640, 1697745622500414, 1700892710022249,
-  1702551278153893, 1704120235453215, 1706176146587689, 1708781142358771,
-  1709237518448403, 1710244072654747, 1721189195348368, 1723648960400183,
-  1723724741017180, 1726128160100176, 1726279575870000, 1726455205636363,
-  1730863672270239, 1734984990617831, 1736545187351178, 1744417312913478,
-  1744959345874305, 1748073142265516, 1752869360078691, 1753741890743970,
-  1758909701473842, 1763202257318778, 1764656756484811, 1767066091372403,
-  1770143836877788, 1776509602212960, 1777501091607633, 1777713688305912,
-  1778015469891961, 1778405387305300, 1779576656587333, 1779903873085066,
-  1780294098549299, 1781103055882256, 1782068211679098, 1787385060086626,
-  1791010333456983, 1794819645879983, 1799093790652187, 1799576109135037,
-  1800844692448810, 1805152492593834, 1805923586789008, 1806131113277691,
-  1807902504763685, 1812859695137579, 1813294737677617, 1813366107002276,
-  1826413364748005, 1826669030844554, 1827688531222649, 1828161134441797,
-  1832031834892105, 1837746644929990, 1840723171692401, 1843215944857316,
-  1849031665489049, 1850407507219180, 1850631764552556, 1855813782771906,
-  1859109335501435, 1863598076685300, 1865674718993878, 1866711341123556,
-  1869317824117879, 1873806874465154, 1874393071460187, 1874745471595953,
-  1877008938281631, 1880472385046063, 1885634820732984, 1890311614182987,
-  1892237778880638, 1893695925388371, 1903577273929003, 1904159424066439,
-  1908358426141003, 1909897941820874, 1909912407845272, 1911689297977349,
-  1914569760053776, 1923320253660518, 1929006184642625, 1929638594048481,
-  1930114947026937, 1932970291330698, 1933095895424781, 1933753853127437,
-  1941279680127051, 1943072771294028, 1945647383634808, 1947788527087377,
-  1951946405387852, 1958021995626462, 1962324291207937, 1969605282327673,
-  1969812476076884, 1969895810529844, 1970754910417400, 1975884269357951,
-  1977629715164696, 1983175522834803, 1983318412071279, 1985295950372952,
-  1986786598339611, 1992837097758451, 1992977322701391, 2004400689125008,
-  2008279712939810, 2019330515352104, 2022024027194232, 2026331820751059,
-  2026689700851266, 2028510886623702, 2030836132688442, 2032913967257740,
-  2034576379866813, 2042765859168900, 2043562846616261, 2044829199441195,
-  2047685495488427, 2050715641288281, 2053482815013965, 2055954462907889,
-  2058176307747332, 2060366930685364, 2064575882010721, 2068541703734446,
-  2070586993533337, 2071651146437919, 2077675523518396, 2081655551912062,
-  2084294928403607, 2089997628026046, 2094070526944895, 2096461769472870,
-  2103985839634248, 2104850026552838, 2113444967197374, 2115229398555980,
-  2118389787254175, 2122488302086812, 2126781234541365, 2127070348763587,
-  2128708515614679, 2130329022405747, 2134286993963884, 2136546685362316,
-  2140156544950027, 2147443799177103, 2161351029781630, 2162646310045907,
-  2164847576666931, 2170735163714206, 2171046709073629, 2176260018094688,
-  2178311572227029, 2181575634966214, 2186747041891826, 2186828324120532,
-  2195115173058157, 2198094050857889, 2198372464090649, 2201658036699941,
-  2204500006490753, 2214072183436175, 2214617197532919, 2215465042239497,
-  2221007330962583, 2221199340037462, 2228997792729738, 2232270103328722,
-  2239271795885790, 2245411562552885, 2246891262703153, 2247679914765485,
-  2249055216698230, 2251233614838571, 2256375296751772, 2256838671564855,
-  2258414749736009, 2262017689403149, 2263651914084571, 2264804977249333,
-  2270138052416303, 2273184505262865, 2273215608469034, 2276146544180592,
-  2277480505851915, 2278640267718745, 2281304274944393, 2286252923358395,
-  2286930275846850, 2287010367601360, 2291133064827126, 2291429211503811,
-  2295826300792568, 2304733564548753, 2308700301228523, 2322743421311647,
-  2325620821689398, 2334410460776075, 2340601074234497, 2342283409009161,
-  2345184292297652, 2345230057834427, 2349868714214804, 2368314149726215,
-  2376156901751945, 2376510596454137, 2390331970441992, 2393240734330294,
-  2394708179726937, 2395365760369891, 2395807876909558, 2403906295594701,
-  2404874589781197, 2406044557642144, 2411126198937921, 2411704488193204,
-  2415797794047311, 2423241771763104, 2424422146139762, 2426882232920845,
-  2428212458770707, 2430926201645313, 2435592103630449, 2436615952067313,
-  2436981010615782, 2458546154956215, 2459251184973527, 2462931056197217,
-  2463537190431969, 2465713037010420, 2468252247899922, 2470283112787569,
-  2473035412680820, 2473174658271293, 2479860962225242, 2480070701908012,
-  2480402345771950, 2485420580036148, 2485736908678697, 2486712527740547,
-  2489040222289815, 2493830660421566, 2497306887910129, 2497636424762009,
-  2499331179256174, 2503549852666969, 2504059296325491, 2506846456775480,
-  2510893746291813, 2511056363055147, 2514542046600264, 2517147620784879,
-  2519606879939592, 2521156984957632, 2526070493995269, 2529395575922913,
-  2529676218824785, 2530819822147774, 2534542888042610, 2536202916887166,
-  2545633830778714, 2546705155460291, 2546851013942190, 2550618099756501,
-  2557375383277742, 2559338753777364, 2562040432935871, 2567146864361055,
-  2567893403112811, 2569531608397355, 2578707080591349, 2580775143486714,
-  2586405089853350, 2592221680479393, 2595531417399606, 2601317770484684,
-  2604560226168595, 2607133397725276, 2611150219742775, 2614710419346485,
-  2617752157584166, 2619545290412143, 2620307613859860, 2623911967221170,
-  2624903755061945, 2632317804036415, 2640721308837834, 2644180844345872,
-  2644579733037027, 2650105405887094, 2656117104453611, 2665947089003486,
-  2669128638985173, 2671539619440888, 2675150905608835, 2676050255302751,
-  2682390499273230, 2692404459752311, 2695078264876543, 2698640393439149,
-  2718304575712694, 2724520124645726, 2725089844754172, 2731610343565665,
-  2731833681954799, 2731978002219877, 2737131191032224, 2742460355647934,
-  2742511833283824, 2743084628266265, 2748984664040514, 2752651065967943,
-  2760875907267484, 2765107755764107, 2765605274681565, 2767985871434802,
-  2768235433305523, 2768809964685205, 2769623545162557, 2770361740671397,
-  2773534910949869, 2776976768586124, 2784164744395598, 2787820568852118,
-  2791164672544774, 2793605516725495, 2793871185415788, 2801576866086472,
-  2802817799710414, 2812088271811913, 2812232885304908, 2812252769219154,
-  2812357296019027, 2812765850303079, 2818329578857423, 2818570055845772,
-  2821487605798285, 2822971164449565, 2823172399624446, 2826823829901649,
-  2827467287751795, 2834506105275783, 2836720740139549, 2843194592867171,
-  2844512204706201, 2845674951089094, 2857921504027848, 2857972489870673,
-  2860802359976410, 2865602104479864, 2867159550353456, 2869683107949573,
-  2878080328717284, 2878212984512316, 2880219009415973, 2883647131829845,
-  2885251274431733, 2885329549604271, 2886103084968633, 2903341553598547,
-  2904354357206201, 2908248009657684, 2913005270211896, 2913079780020252,
-  2913438800578617, 2913494541533656, 2918120708680137, 2920330402723703,
-  2922796529630030, 2924769128911332, 2925837935996904, 2929015237514367,
-  2933755868081781, 2933979382564762, 2936030201231965, 2939971299420101,
-  2941375228989512, 2943675880336088, 2947010015531481, 2956552216398178,
-  2957120186366112, 2960867061506739, 2966222289084920, 2970145566916660,
-  2971548914562607, 2972473106994005, 2978118039821586, 2984205870899483,
-  2985062270777047, 2986349395050842, 2993430959662262, 2997929996196299,
-  2998028226543625, 2998504017695283, 2999743208983350, 3004444071022789,
-  3010910449509393, 3011737329480745, 3045362495387037, 3048673495712182,
-  3053044088315971, 3053062590432128, 3058149371560096, 3058545666955585,
-  3058827788614861, 3063628295570889, 3068624241283024, 3070351273558271,
-  3072766041263354, 3074344072991099, 3075171041334179, 3082226652293577,
-  3096124608323736, 3101807954679829, 3105890702218611, 3107245298914857,
-  3112679603951766, 3119197705838897, 3123361853664708, 3128249227386574,
-  3131644092925694, 3131689855894803, 3133172126399264, 3133583203444376,
-  3159839492383058, 3159916225013213, 3166614976721514, 3167402486946730,
-  3167945828649919, 3170946408016138, 3172565331808711, 3175085977501336,
-  3175895757975610, 3179761479740352, 3180392530100142, 3180543269697414,
-  3184449891556348, 3187099323346858, 3189688629392979, 3190844064736976,
-  3204962911528959, 3211271897909699, 3211360638736235, 3212243373099017,
-  3212519258528478, 3225026907169718, 3227226756634792, 3230898439516359,
-  3240029832281746, 3249675177811140, 3252664789103184, 3255363132263390,
-  3260001071703517, 3264287345650629, 3267700841198121, 3268189858110763,
-  3274889226566330, 3275687392376996, 3279042428544752, 3279584487474843,
-  3287607950213474, 3290992481514931, 3297461130568727, 3299765354804022,
-  3308223948232064, 3310068802998092, 3313455434108562, 3317753267484041,
-  3318161371366040, 3323872738617697, 3324964102973517, 3327154403117147,
-  3339558127051513, 3340741665619248, 3341021177771397, 3342736902553276,
-  3344352593492104, 3345004985818262, 3351124734230955, 3352472768520681,
-  3354035615313459, 3354834429027294, 3356052754749783, 3370260866007062,
-  3370548400622315, 3379214183488034, 3389730943133671, 3395603784898046,
-  3395649472091180, 3396547802484157, 3402818317576466, 3405578276197731,
-  3407782769926759, 3409354104219916, 3410161829057290, 3410814963473334,
-  3412127647807997, 3415194875598262, 3416358509423267, 3417068746387064,
-  3420162010525255, 3422105908061587, 3422361667496604, 3427344107563170,
-  3429309838799524, 3429722859214785, 3433159884643763, 3434854934212629,
-  3435218010812643, 3437456904084814, 3442480450557212, 3442927759953117,
-  3465753258024557, 3467759507688093, 3474663003300773, 3486681496115438,
-  3507290020750642, 3508501827263918, 3510481791918582, 3510633064845919,
-  3518226470377086, 3525456469383042, 3528779219038091, 3530846349530505,
-  3532310623144728, 3533516080490270, 3533716524311972, 3536808666815407,
-  3541481510805829, 3542231526260195, 3550852997316471, 3551457842836964,
-  3551530209564137, 3563192956089398, 3565495442464095, 3569248623677618,
-  3578078519389609, 3585584910754392, 3593755765431323, 3596131140602191,
-  3604939621075607, 3606058898547232, 3606146277306109, 3610525955882303,
-  3613079942175374, 3613625671810412, 3627368021937366, 3628971798264251,
-  3634753212886267, 3644391848644249, 3652215957133607, 3654244364977988,
-  3654300102808886, 3655209934092358, 3655569814914594, 3660106161174706,
-  3663361817260580, 3664560735942173, 3666918801844311, 3669087314812690,
-  3670539104837876, 3681653015228264, 3684589603989399, 3697673821383339,
-  3698150275855942, 3698796632593010, 3699452753596952, 3699581543356780,
-  3702627178845230, 3703406551735408, 3708467189866567, 3710091724644859,
-  3712374942456149, 3716551007073465, 3716790539331183, 3718419844978808,
-  3718570777177101, 3726631885159258, 3734426556555842, 3737320332094012,
-  3743536725338210, 3753894917367563, 3755010366083299, 3760758873165303,
-  3767495798103370, 3767713766482457, 3771891827926166, 3772145736370861,
-  3774972460085576, 3775233272004140, 3785255994134362, 3799815315012454,
-  3805951868843714, 3806848795938769, 3809231504672393, 3817695860383153,
-  3824857247377129, 3829302190006219, 3830877204298875, 3832576323776219,
-  3835941601795248, 3839715369925076, 3846670068125361, 3849200137952497,
-  3853657556921710, 3858119166993230, 3859165718613315, 3860279009384183,
-  3861645575185022, 3861873906764950, 3864166368559083, 3869969942979283,
-  3872868533635323, 3876948058807299, 3880412057105785, 3886485814537147,
-  3887015258810651, 3889926103905770, 3896374551556710, 3900397325715887,
-  3903975186892728, 3907596869800719, 3908640663442050, 3913052772369566,
-  3914173271671437, 3917568103678165, 3921202743566465, 3924388393712595,
-  3927183846580591, 3929470500322313, 3936847289209131, 3940132980172322,
-  3945881211033660, 3950429752758333, 3960742476534474, 3961966437522729,
-  3964632496801597, 3966258146665592, 3971007665815295, 3973312312829630,
-  3975323027734760, 3979411501232208, 3981302967516837, 3981353374167011,
-  3981981596300992, 3982654043363879, 3984023182452270, 3991585129970208,
-  3993322362615082, 3995587717731273, 3995695025325733, 4004678361890085,
-  4008001717746910, 4019103937849892, 4019987284179295, 4021303360540009,
-  4022227942295604, 4033166066402694, 4035223922084504, 4038648846025435,
-  4042540791312058, 4043124706409454, 4052302230031797, 4054072183393722,
-  4057483474995143, 4062053943055931, 4064567574139564, 4069773402555351,
-  4077944369138857, 4079752922216011, 4082376245036300, 4092882813566934,
-  4094060189043417, 4094217937327192, 4098654295034081, 4099621987536298,
-  4112400091106811, 4112957866069707, 4114291113080623, 4114508630118504,
-  4114776343526748, 4118019629018797, 4120793713070614, 4124966250988201,
-  4134516711171537, 4157334505825170, 4157900443891044, 4167776152389018,
-  4168499961321507, 4172587549868852, 4178059958670376, 4178383739372395,
-  4182088947283013, 4182946140136623, 4184851088105735, 4187282042893941,
-  4194845928624819, 4197492683303412, 4199112897285624, 4203386020622584,
-  4207294456422787, 4214507651111735, 4223537598235888, 4227618581508497,
-  4228196909958910, 4228624657943598, 4228832806098027, 4232147444444168,
-  4232900020503464, 4233171144314311, 4244106383316547, 4245867296460237,
-  4246930437629566, 4249826160427021, 4255544580546916, 4257590677706052,
-  4257605711945981, 4260226929046691, 4262429800654049, 4265070701827248,
-  4265998406131300, 4266082875232996, 4266226535122919, 4270255297284740,
-  4270865098680487, 4273369687280598, 4274273088847614, 4274836575607892,
-  4276989633478281, 4280759321877343, 4282858550337928, 4283893263439354,
-  4288163822749037, 4295518888361560, 4315560235236611, 4317167460095430,
-  4317173734230465, 4321443682744041, 4323711084349085, 4328777675124169,
-  4330092906304499, 4332285863753267, 4332446524376244, 4334963796716111,
-  4335023700003100, 4335220668045715, 4335321460782402, 4337248664637506,
-  4338959116832859, 4341678395557291, 4347903940514340, 4352173337770984,
-  4357200597246264, 4361602523614552, 4362484954310088, 4362986574641945,
-  4363204944265029, 4363734400768130, 4384395784655077, 4385310167003174,
-  4385493681841220, 4387743619419313, 4395249475257937, 4402125969562283,
-  4413064627744832, 4417790078900339, 4419126035107133, 4425458055521407,
-  4426125221337959, 4431918903234859, 4437448444823850, 4437778312172988,
-  4439049044178904, 4446144485258736, 4447392920302285, 4456659336515543,
-  4459118321877640, 4461941621292855, 4463206774842489, 4467872852230954,
-  4469363012983105, 4470934156859601, 4473810290942350, 4474094950211202,
-  4474176406766793, 4479888499744113, 4481121569832617, 4482379867407668,
-  4482409833594925, 4485436794139625, 4489787200174406, 4489975886763358,
-  4492082706234236, 4493669513583525, 4496673333070260, 4506849614290634,
-  4509547136804989, 4511805659766186, 4515349774347434, 4518733511439826,
-  4524092177562783, 4525871612082560, 4536628070345079, 4538099859745215,
-  4541552317353976, 4542551474492183, 4545542879617172, 4546028005876506,
-  4546122388516236, 4550041696184369, 4550302275492245, 4552080957360830,
-  4553317631132795, 4555106566429153, 4558158696143217, 4575499549522017,
-  4578802988070000, 4579981126888956, 4584828011313204, 4592841496735960,
-  4595692197433149, 4596537241184602, 4598507276516208, 4602187272081614,
-  4602930520840431, 4610647977694781, 4613868682609971, 4621915763432984,
-  4622441005133054, 4623153438679712, 4624513901814664, 4628571765184899,
-  4629458252343533, 4632904172757003, 4641429354859666, 4644245195166272,
-  4650611639707243, 4653196328024765, 4656668338048112, 4657844180731782,
-  4658881017820748, 4663222826965654, 4663308591654074, 4663405540693671,
-  4671209434741542, 4672323647200368, 4673964836077761, 4681841554933008,
-  4683847747655569, 4684676697969273, 4685357244727722, 4686038374465449,
-  4687085319186935, 4687434118474408, 4688420683590210, 4694767695342162,
-  4700504115908871, 4707273151169133, 4713365854017461, 4717681269221088,
-  4723913819744323, 4723946132504426, 4728741981175152, 4729416627452139,
-  4732812854015886, 4737774501119146, 4738353474108524, 4746403105415883,
-  4749462418300841, 4757995478658971, 4764048112863023, 4768288894490020,
-  4781758900783499, 4782079619989332, 4783978968293976, 4785296059565013,
-  4787375451033414, 4788316710394849, 4796353377509258, 4799317070670822,
-  4799615622826170, 4801630387788470, 4803605729664023, 4809617770077414,
-  4823460388124897, 4823567256367860, 4830199860604563, 4837902092503579,
-  4847231138477573, 4851929144628376, 4854108117029869, 4858290146627622,
-  4859819010105941, 4863477069218996, 4863612129453418, 4863798729157421,
-  4867846704623585, 4868694874124956, 4871613208043515, 4875265795491699,
-  4876087570860731, 4879641253729269, 4881872310639574, 4882057576668537,
-  4882743535486975, 4883341047127586, 4884656261897757, 4886100305047914,
-  4890262920236021, 4891448056682821, 4893322677066788, 4893454989124078,
-  4896895411828937, 4902998512973858, 4903102728788730, 4903941125922973,
-  4907752183874047, 4908579472146469, 4910269376997752, 4911978397265523,
-  4916434931602774, 4924165317523917, 4927490848989786, 4929808494513772,
-  4930877891617749, 4931140461756538, 4935049000174841, 4935054907554937,
-  4938200090949099, 4938585085595100, 4941097090785636, 4942944526012603,
-  4944871865363172, 4945921416930566, 4946012814181195, 4952361461467789,
-  4952691731251188, 4958653076583442, 4959752681080291, 4960896820999356,
-  4962041133447084, 4963673350125525, 4963945258961417, 4964814136043531,
-  4970280875161319, 4973982849282023, 4976153646284424, 4977451144311263,
-  4978433039301078, 4981020320288919, 4982699388214246, 4983437343560821,
-  4985223644218718, 4985516623582081, 4986346666211613, 4990459068637829,
-  4995559349965340, 4996132778910598, 5003001506302560, 5012186217884698,
-  5014117845946597, 5015906848625987, 5017655250127530, 5017941801840002,
-  5025562651085838, 5025813436191669, 5025969574644644, 5026274448291818,
-  5027184215525036, 5032049075769790, 5033278563408545, 5034601143759093,
-  5036961073226369, 5040680239641462, 5051698431429311, 5052005680885739,
-  5055368359149551, 5056710463050956, 5062909013629968, 5065614434582112,
-  5073479171670861, 5075636765877202, 5078312917977063, 5096132068319894,
-  5097391883235138, 5100509442146956, 5107435285811708, 5120630462488949,
-  5121288122705812, 5130157433652247, 5130997309601188, 5132856450888714,
-  5133768083871538, 5134065548613979, 5141126769907420, 5145518209772356,
-  5145938191779447, 5147952091782267, 5148694611273651, 5151165148653410,
-  5154015712463142, 5162482736617575, 5166433445011092, 5182211629862944,
-  5183440891577798, 5185042594869499, 5194409546432841, 5201635247489017,
-  5207769127133715, 5209313464077735, 5210900995400634, 5216341942629215,
-  5217146767840924, 5217383462667847, 5226579728276621, 5230392153332044,
-  5231489194595314, 5232423259863162, 5238120075808996, 5239332795048450,
-  5249535530653482, 5251433751251853, 5251888375535870, 5256166520825884,
-  5260701908170551, 5261638505507060, 5265993490824007, 5271306384741574,
-  5271415702830127, 5277301708461683, 5278436243540888, 5278888406088488,
-  5279938600732187, 5280160783203712, 5287473234654617, 5287724161264034,
-  5288599148476195, 5289080363207105, 5291899306749784, 5292622198185453,
-  5295942629716263, 5296642807816477, 5299472296897255, 5311802217264885,
-  5315721237515608, 5317267291655431, 5320546962146697, 5326186663925106,
-  5332901445491331, 5333661045292905, 5333900425748672, 5338914689666091,
-  5340585230396366, 5341536043202386, 5344017800803446, 5347521869218614,
-  5349298160474328, 5353981582803452, 5354575331481678, 5354679095326160,
-  5357678939079853, 5363410349016514, 5368862525078129, 5369674679314474,
-  5374625409868960, 5374920511387205, 5375379411293931, 5379875453259544,
-  5380132808287800, 5383865794274589, 5387630299951409, 5394175358493646,
-  5401237579120936, 5401553262018904, 5403263460692404, 5405249976671550,
-  5407675007222893, 5409319022542673, 5413507006148389, 5413922933902304,
-  5414455243531978, 5414562377248070, 5416142154655251, 5416162557239732,
-  5416168517430092, 5428299398099562, 5434029123414897, 5435643623336163,
-  5438571225587456, 5438651755404083, 5444321287105819, 5452694597139958,
-  5458646602054625, 5464807475383067, 5466227086057958, 5466540270936040,
-  5468890988865976, 5472650576605613, 5477785240360047, 5479661009477987,
-  5482454652787382, 5485479613743962, 5486715083181383, 5488141709724584,
-  5489430347488044, 5491065222692103, 5491226079126096, 5491797026361768,
-  5492015428770431, 5497092831873764, 5497541657115384, 5498260126036916,
-  5510709356292792, 5512685456225112, 5512866704444110, 5522023931338073,
-  5522267973293137, 5524266979258349, 5528988425691027, 5543696559433090,
-  5547888695303249, 5551494023089853, 5552234994271139, 5556118190379020,
-  5563398553948901, 5568517576242237, 5575878127955354, 5577802937911513,
-  5578147869037370, 5580639527663640, 5580840739203419, 5582176681183982,
-  5584660581442677, 5586788283932281, 5587994733177356, 5590214652991215,
-  5592410166429812, 5595475014807940, 5596397856778238, 5605899595134065,
-  5615976985450312, 5619275859233761, 5622078880641017, 5622444440585394,
-  5626949796077003, 5628062637334907, 5628610184020193, 5634560095333076,
-  5638122116649604, 5638529384451970, 5640982942480838, 5643526074631722,
-  5646299014463606, 5652934248029101, 5653972551965764, 5663707381902353,
-  5665839271389087, 5667041696390327, 5671120475279840, 5672485589396838,
-  5687597369379625, 5689386437065026, 5689641062490265, 5690971266446719,
-  5695868006279620, 5713810325038812, 5714110934034480, 5716380546768346,
-  5719573838462785, 5723448462041614, 5733626020403711, 5736571244245678,
-  5736732954877112, 5744649432909271, 5751486951885446, 5761266907592326,
-  5762463462550868, 5763339261777587, 5764115091215928, 5765048977871492,
-  5766556448259849, 5770864103514011, 5777287227161577, 5778902107995447,
-  5779047677910458, 5779760889029436, 5780424446399056, 5781222988256846,
-  5785437979966298, 5786257548529041, 5790424311486242, 5794130664267168,
-  5798766868342238, 5799672864989698, 5802860738758769, 5805708251005077,
-  5808786529336006, 5810424320221552, 5812824942344676, 5813026598102548,
-  5814024874391031, 5816763610753650, 5817317318043100, 5819257047206331,
-  5823840772407985, 5824184178955623, 5831978101114534, 5833183465698407,
-  5834211249917848, 5836966745254795, 5847960743524350, 5857529669296864,
-  5860790296191183, 5861950483943708, 5866068066241247, 5878241407330590,
-  5882994861887693, 5884258533925519, 5885082165577428, 5886355407109785,
-  5886766674928943, 5888927363157049, 5889548450617598, 5890059430549506,
-  5890494539744317, 5891726836013141, 5892410623868647, 5892992473916987,
-  5895989815568943, 5902435713234829, 5902899115704417, 5907423354195347,
-  5907505512493912, 5914537734882196, 5917195732538035, 5919829311288137,
-  5922367331491744, 5927349298289295, 5930007137225413, 5930451700440843,
-  5938857839397052, 5941861009891431, 5942588650175719, 5944434134179133,
-  5944932205750164, 5949118180125335, 5952729429183626, 5955135933961818,
-  5958091429563596, 5959509406206871, 5969003580018688, 5975654829160753,
-  5978097664714324, 5984491605180390, 5992060361861373, 5993734326764317,
-  5994273326614838, 5994789073480801, 5996167995970074, 5996296762005840,
-  5996482505638952, 5996660942948155, 6001996212954131, 6002608437371545,
-  6010513842474496, 6011745048862095, 6012963526266495, 6018151997343975,
-  6026663245013264, 6041840082587001, 6042974982011406, 6044083939030347,
-  6049033493828838, 6049432580287298, 6051818476521374, 6056740170378647,
-  6059314108184256, 6061899932754560, 6066760566987649, 6067148412526389,
-  6071409502527587, 6076082332551187, 6076603819866530, 6079307959029706,
-  6084783460037724, 6086952731513098, 6087238476836324, 6093830274323943,
-  6094422985191305, 6099960458338468, 6100320886081913, 6112789954005368,
-  6113070386900554, 6113669500226969, 6117744809201616, 6118342019824793,
-  6123831921937391, 6125332142798587, 6125603760517113, 6129979655355093,
-  6131555275488014, 6134549288973988, 6139132662780799, 6140489271086000,
-  6142967060280622, 6143888909718360, 6144069089622740, 6146323283272506,
-  6149426607619198, 6151803516346010, 6152025153107183, 6155305119122198,
-  6159420228314507, 6164368449231552, 6170554380812211, 6174371361896883,
-  6175860249032388, 6178171992461941, 6179254306802956, 6179562834672238,
-  6179755909970360, 6182085494898038, 6182245452368640, 6182336345315594,
-  6186717198817324, 6193611462300413, 6194436840473847, 6196862838553754,
-  6198291055512338, 6199845894247618, 6207033359680251, 6210623623008755,
-  6215495431995062, 6221253018532949, 6227064139227087, 6227903257714994,
-  6228026142537823, 6230763228107878, 6241292103300816, 6242422960448593,
-  6260321858983918, 6263000397734357, 6273097395593245, 6277364345107404,
-  6280055021013025, 6286427801313522, 6287488184564254, 6289244298242100,
-  6289874977064087, 6290046225906214, 6292395433687434, 6293058521869172,
-  6296958298250958, 6300531071864648, 6301021182831013, 6306840231680425,
-  6309300307266476, 6312886760898213, 6313769499851601, 6317510396810434,
-  6318026718003481, 6321426590663265, 6322364564071035, 6323998920784681,
-  6330210666044079, 6343383023176497, 6344296598035568, 6346618432151421,
-  6348938957906295, 6353152919725603, 6364722312338366, 6378698547569866,
-  6380373852229220, 6381716357846472, 6391210716262122, 6396044616451768,
-  6409851286013877, 6418194110370323, 6421958473176463, 6431348884329004,
-  6431728681187460, 6442490218371299, 6446140856979113, 6449714925430838,
-  6450911293971600, 6451337427394816, 6454699095307099, 6455465856527962,
-  6457640988230668, 6458267402376516, 6465218405840151, 6468013763772164,
-  6469105356993349, 6475530061295179, 6478139316429842, 6482114766137666,
-  6482755832557680, 6482941304047546, 6485211449999241, 6490691373466543,
-  6491350383019651, 6493367441179748, 6498880486857635, 6504510317065310,
-  6512448276900380, 6513472977316171, 6513605988947304, 6517451556868444,
-  6520112168581173, 6524483353515134, 6529588775394100, 6529647155791724,
-  6533353234513649, 6540087834964530, 6543855537785199, 6544771655455125,
-  6551534053851888, 6553153034262975, 6561240416352655, 6563729376575033,
-  6563836912577572, 6566557458316615, 6570828848562495, 6578659812361884,
-  6579148356216462, 6584621498208478, 6589046919061309, 6593869546772204,
-  6594097167228997, 6601273674115918, 6606866654629945, 6608166546472864,
-  6615980124509252, 6618238262028423, 6621911771130261, 6624666507967854,
-  6629009834907846, 6635258775706294, 6638482986177208, 6641285251852398,
-  6648791609237767, 6662036075565964, 6662710687286213, 6663445720936926,
-  6663707638510483, 6671361328833595, 6675297389667116, 6675835105027290,
-  6676977379105582, 6679279433898202, 6683991837542818, 6684126697660290,
-  6685923079162940, 6687042789888737, 6688898497782576, 6691325963768703,
-  6691802555595822, 6694952694187413, 6703201124799689, 6707788806564806,
-  6708017955711463, 6709625067254490, 6710927014045529, 6715102534094077,
-  6733014110297302, 6736226359501915, 6738892380516831, 6739155857203715,
-  6739644468908876, 6739946166269733, 6745268510327484, 6750042770069277,
-  6757335880582957, 6758573536092466, 6762914492951347, 6766298734037888,
-  6767585928288493, 6771568703496798, 6780621504871415, 6788380548462288,
-  6792845369799125, 6794036230450459, 6798605994894845, 6798706568572195,
-  6801031533190571, 6803144096011317, 6804419688423692, 6804625562307193,
-  6818552418540543, 6821207209198830, 6823660949610012, 6826774918530594,
-  6827240692079051, 6828810442752318, 6832395631195652, 6849438146211848,
-  6851087692456894, 6862653686094552, 6865836301257369, 6866172721594213,
-  6869336637866936, 6870999575375660, 6872014145946187, 6880851063868033,
-  6881819857828472, 6884761636184597, 6886888478678924, 6887424937289863,
-  6888687046708838, 6889931603171473, 6892427314910687, 6894771254842458,
-  6896773855143484, 6903092902856611, 6904880510445361, 6910026135653077,
-  6910894485604843, 6914875534862612, 6917932710572611, 6920622430798682,
-  6933984059850961, 6934740874010750, 6934862763558194, 6939401751960061,
-  6941502934858942, 6941996429641337, 6943296865636872, 6945876423757611,
-  6951883680487612, 6952972818604603, 6953976388572781, 6958372348348893,
-  6961202773960681, 6966068430383674, 6968969608315193, 6980690024780875,
-  6985747933814622, 6987097769746084, 6996025070029249, 7006834280876281,
-  7014309621033759, 7022505815866595, 7024505881409531, 7024897021133738,
-  7032308987562704, 7047067617181295, 7058796324176673, 7063237739729606,
-  7063735889097181, 7072933372232853, 7074773743423137, 7075738961934704,
-  7075900527233804, 7076700314221798, 7088851360303745, 7090884269120417,
-  7094709670546306, 7099675635683327, 7107618666169985, 7108926506516470,
-  7110867305325040, 7110883513238170, 7114416375258716, 7120925585747960,
-  7129054027188848, 7130360392504062, 7132286054468357, 7139849171278012,
-  7141554258270033, 7144684521617112, 7147638547793137, 7158284286725925,
-  7173410056743859, 7174776873979868, 7176456894067982, 7177390538352051,
-  7188584474897378, 7195606292633615, 7212109147401252, 7215524145689081,
-  7216046767295134, 7216182325933228, 7216467190801240, 7222706570544202,
-  7224718475300993, 7225876513432313, 7228295075319639, 7230219592890071,
-  7230775807716758, 7249374643094015, 7254477981062528, 7263105934293265,
-  7266787909800913, 7267290524561122, 7267303850780530, 7268112841425950,
-  7269584934916617, 7274900560716876, 7275894844441878, 7278671879778348,
-  7279353916579668, 7286590607677936, 7289123791598156, 7292590660544294,
-  7293758923669493, 7295440661251781, 7297388387610172, 7298220730040217,
-  7300529645880188, 7301875026042253, 7304076277125547, 7305158389913400,
-  7312377365942862, 7313230518604709, 7314977313998828, 7315087799023784,
-  7316054912814337, 7318994913462659, 7327528965873867, 7328013597245938,
-  7330468129402942, 7333418988608969, 7335807019654874, 7338850268573136,
-  7339951949996884, 7350008068959815, 7353181391780741, 7355778780626783,
-  7360475559368956, 7364144384681590, 7364714206715255, 7365228869775600,
-  7366793369594098, 7368020891231175, 7370610251281497, 7376456096414683,
-  7379735030644043, 7382010546552564, 7382687800330635, 7386057343962975,
-  7392727366044467, 7402845453168332, 7408940190918604, 7410614271653986,
-  7411069685111255, 7413105879109452, 7416402508905762, 7419002859612613,
-  7420519350736091, 7421258074806310, 7423430625081601, 7424077539478697,
-  7424233213398831, 7435544494485487, 7447538268692242, 7457522703386213,
-  7457569083486942, 7459344003921524, 7462851611192235, 7465785070437250,
-  7469105981962405, 7472571780125860, 7474529276271610, 7480805138803249,
-  7490148512361838, 7496715025460605, 7505531161704911, 7505671213725642,
-  7510568952022088, 7511089644397807, 7512946562776188, 7513002782962582,
-  7517186310048965, 7517710332882783, 7521334587738672, 7525629337372211,
-  7526559844968696, 7532585420040345, 7533397794794217, 7534129889863032,
-  7538560123407679, 7541270842935629, 7543028875556438, 7544449234690402,
-  7555667330233583, 7558324245570973, 7560560872747199, 7562084073235262,
-  7562135592877826, 7566883551048359, 7572098114760836, 7576108381251833,
-  7579675214793854, 7580681941096029, 7580838132039731, 7583703348643460,
-  7583814537877607, 7590521720630198, 7606174043253607, 7609250439941817,
-  7610439419764857, 7615836642129235, 7618093386314088, 7620691247554374,
-  7625297673408058, 7626232026079928, 7628716817324694, 7635826958285406,
-  7636100908665817, 7641613018116162, 7642108803564713, 7643798277586412,
-  7645241714110996, 7646703602634222, 7646825092751898, 7649595025255347,
-  7650177869329937, 7651107174379887, 7652842740271078, 7655553280254272,
-  7661632370266221, 7662121822688543, 7676954292933978, 7677443931979443,
-  7678368406582782, 7680490791858914, 7682575905726058, 7687072275490272,
-  7690374473937288, 7691950882418698, 7698963790452769, 7700488573158414,
-  7703271270737362, 7710897059199750, 7711412579978379, 7717200128102908,
-  7720351978951940, 7722926445523199, 7727887735435983, 7729966068821404,
-  7730501966635643, 7732868269840659, 7734576229583529, 7735907671179130,
-  7736594464017600, 7736889985806069, 7736904737851329, 7747343835511086,
-  7754479525873539, 7758181388463768, 7758547347648222, 7758715002172683,
-  7761353611533412, 7761715688076258, 7769582597045005, 7772485870799785,
-  7779870450913252, 7787616976572041, 7793496162257462, 7793885993713752,
-  7796769673839041, 7797766887055176, 7799591755820645, 7807010441720273,
-  7807514156201632, 7810222244835374, 7812469485694152, 7812475487729565,
-  7815503515707841, 7815984945470097, 7816874139002453, 7824191698834760,
-  7826880247102186, 7831549314999317, 7837034328394180, 7837105249138606,
-  7842675309738413, 7842813344634660, 7844535269713322, 7844986565491916,
-  7849590413042820, 7856898589341882, 7857398688506942, 7875697994982526,
-  7875757663254442, 7876037604210636, 7880874159364410, 7887205619332695,
-  7890657516237149, 7892125882909070, 7894353496788525, 7894426922979657,
-  7895992368809400, 7897500402050036, 7897635924063980, 7901451450115544,
-  7904024296657231, 7905142593353563, 7910692800197392, 7915854898398868,
-  7919025593010668, 7921061254720266, 7932922717082023, 7935366695384815,
-  7936292824450488, 7937809658778315, 7941035279216096, 7943719466236709,
-  7944305289662981, 7963278312960991, 7964406972323236, 7966216462706218,
-  7967987367091018, 7974361123100741, 7975434993463626, 7977690821598628,
-  7980169320458408, 7981533128387239, 7982936463721611, 7988530855828939,
-  7989835217839853, 7994669975336686, 7998045002581906, 7998718799664178,
-  8013172716699246, 8016207989050117, 8019027548852313, 8023390527579661,
-  8023586213391790, 8032114176355142, 8034101681646748, 8037009789258727,
-  8044921704375888, 8049664944022401, 8053403929014542, 8054878482888818,
-  8056973986994274, 8057999291682258, 8064606307728906, 8066289729554241,
-  8067174063268292, 8072606616631997, 8076266729681341, 8077645829582682,
-  8079723692045693, 8081871720099818, 8083979865478945, 8088432089644539,
-  8089603672978061, 8099155962755249, 8101556486440684, 8109585531183676,
-  8110953911473600, 8112158670984405, 8112921925061116, 8120622565861391,
-  8125472792346100, 8128648950180250, 8133822563090483, 8135895591413011,
-  8139672593514929, 8143332169335919, 8144884182209257, 8145597990164506,
-  8158259459714898, 8159801992704935, 8163105492946124, 8164711074080915,
-  8172481307281306, 8173849792927699, 8174617130757426, 8174724876676662,
-  8180527321196813, 8182508560974213, 8188191463893676, 8189205630581139,
-  8192917292825473, 8193323419745618, 8194608849080847, 8198080742681953,
-  8198469760519926, 8198968431511048, 8199386928101616, 8200741495504176,
-  8204019245734281, 8213217532397968, 8217113170640945, 8223820503747219,
-  8228632607357396, 8230232038770084, 8230492386297364, 8233995591573276,
-  8234969399314117, 8235316087591345, 8235448659261417, 8237676590827473,
-  8242211022949519, 8243711493457744, 8246118938151452, 8252028681142554,
-  8256304913454671, 8272605145388332, 8279650199703967, 8285835292359108,
-  8288216890608527, 8290657758710783, 8291081813553969, 8294449174256879,
-  8294590408036157, 8302107272641597, 8303843117807872, 8312454650309873,
-  8324893063928786, 8332710592263842, 8333896029691490, 8336266157788233,
-  8337054946597531, 8340588678122862, 8346134876021595, 8351558765156722,
-  8354482940460943, 8357433611819895, 8357447487032808, 8360659277793808,
-  8361312365492685, 8363283528901729, 8366442488975204, 8366633633883171,
-  8367432066631981, 8374780754285573, 8380107712655270, 8381168657837683,
-  8381733064609443, 8385823315454274, 8387948227429402, 8388095334876967,
-  8389617235836012, 8390348966674040, 8392365755652021, 8393109671929718,
-  8395450317352598, 8395772494720957, 8398361247730390, 8399238495580414,
-  8404581227072029, 8408543804141445, 8412054509344342, 8413036338447683,
-  8418221770332429, 8419613734687886, 8420195203496470, 8425598694660774,
-  8429101454634705, 8429990637688530, 8431417673437455, 8435916687888788,
-  8443189246533760, 8445695448448615, 8448154068946318, 8449146329204933,
-  8457417310946373, 8460767780066008, 8461492461937682, 8462343776375329,
-  8476852272110751, 8489224874204714, 8489322128868588, 8494622600423216,
-  8500493989199063, 8504062533729952, 8504275704767506, 8512881582649175,
-  8517934848956172, 8527218255660129, 8532286116191601, 8536388352327469,
-  8541715302420731, 8551739578347593, 8551908228985973, 8556076119442728,
-  8557731464550412, 8561488755812962, 8571404678262400, 8573735267923570,
-  8574425772767701, 8576015133052340, 8582035510820106, 8582038389909496,
-  8586781833615266, 8593681026234866, 8594027057735767, 8603455209676138,
-  8604185310381090, 8604496107661813, 8605504932922862, 8607938171830955,
-  8609699460760050, 8618646360218225, 8620679337924877, 8625785399093786,
-  8626322842196122, 8626614134883282, 8627631092075991, 8629339292006787,
-  8638521897238787, 8643691162981039, 8651020188751101, 8651067566290865,
-  8651369559357987, 8654145173893110, 8659384710997355, 8661879543641463,
-  8664390635316711, 8671686204291265, 8673300658804124, 8681741573787202,
-  8685913270984293, 8691833869555204, 8697225546198304, 8698764357879005,
-  8700360787859000, 8701987931165051, 8705003186650736, 8705725572166753,
-  8713726729047476, 8717861607420032, 8723926959931411, 8728385606825882,
-  8728853419082460, 8736665036442022, 8737991936110873, 8739392258972734,
-  8739496424601395, 8742435417164939, 8742683012163451, 8743735977293624,
-  8750606205925921, 8750642297610390, 8755400943714740, 8755813072369491,
-  8756263177398349, 8758462331187694, 8762062104100807, 8764161760499356,
-  8765076136090364, 8767149440026811, 8768180730013995, 8771922417015506,
-  8771992507524811, 8772111532247388, 8774620226557305, 8780162041735952,
-  8784909497507658, 8793682287034627, 8794384265240002, 8803768566424141,
-  8803930560274206, 8810791366031609, 8812320470913861, 8817584789821382,
-  8820608717155900, 8822282204312050, 8824290206091341, 8824722823143847,
-  8830067765375503, 8832222103446407, 8832935128938328, 8839601394762615,
-  8841407747094580, 8845049728990935, 8850452206812949, 8850693033101134,
-  8854630607161239, 8855901193700240, 8856103455883620, 8859188411536517,
-  8859786960013301, 8860184687504865, 8862635414177013, 8863761729893939,
-  8866456582001336, 8874685150837402, 8881246375602759, 8881768206979472,
-  8882542482786830, 8885917897268884, 8890245409430366, 8892547841182470,
-  8894610084055825, 8896904991428579, 8896960011124303, 8902615907735598,
-  8902867212360252, 8904646124096293, 8904714716539672, 8905787981827145,
-  8912803382565129, 8913540752731275, 8914418307961634, 8917339068267703,
-  8922094825546274, 8923115959602651, 8929796828085540, 8930038938308086,
-  8930525781706669, 8930979505344296, 8931392368757941, 8937963893938951,
-  8949120935840502, 8949483597988889, 8950455266245706, 8951137211508277,
-  8952898921068876, 8961469000538202, 8968185550143689, 8977716410631117,
-  8978495456385577, 8980662666495410, 8981105831109515, 8984903494607898,
-  8985134443429015, 8989297873330388, 8991873977440419, 9005819784766789,
-];
