+++ src/features/island/hud/Hud.tsx
@@ -22,7 +22,6 @@ import Decimal from "decimal.js-light";
 import { BuyCurrenciesModal } from "./components/BuyCurrenciesModal";
 import { MachineState } from "features/game/lib/gameMachine";
 import { useSound } from "lib/utils/hooks/useSound";
-import { EmblemAirdropCountdown } from "./EmblemAirdropCountdown";
 import { SpecialEventCountdown } from "./SpecialEventCountdown";
 
 const _farmAddress = (state: MachineState) => state.context.farmAddress;
@@ -165,7 +164,6 @@ const HudComponent: React.FC<{
           }}
         >
           <AuctionCountdown />
-          <EmblemAirdropCountdown />
           <SpecialEventCountdown />
         </div>
